refactor: change to delete empty query
